requirements.txt

cohere
faiss-cpu
numpy
pdfplumber
python-docx
scikit-learn
python-dotenv
